Added and fixed image fetching. TigerDirect parser is tested and working.

+++ backend/transparent/core/ProductID.java
@@ -2,7 +2,7 @@ package transparent.core;
 
 public class ProductID
 {
-    private final int rowId;
+    private final long rowId;
 	private final String moduleProductId;
 
 	public ProductID() {
@@ -10,7 +10,7 @@ public class ProductID
 		this.moduleProductId = null;
 	}
 	
-	public ProductID(int rowId, String moduleId) {
+	public ProductID(long rowId, String moduleId) {
         this.rowId = rowId;
 		this.moduleProductId = moduleId;
 	}
@@ -19,7 +19,7 @@ public class ProductID
 		return moduleProductId;
 	}
 
-    public int getRowId() {
+    public long getRowId() {
         return rowId;
     }
 }
